Extract the duel PvP runtime into a factory, and add the Pokemon schema
Ed, Edd n Eddy RPS and the incoming Pokemon: Fire, Water, Grass! are the
same game underneath: first-to-N, simultaneous blind reveal, three moves in
a cycle, same daily points pool. The plan was to copy the 892-line runtime.
An adversarial review of that plan argued it out and it was right.

The runtime contains SEVEN references to its match table -- six Prisma
delegates and one raw-SQL table name in the award claim -- and every one
fails silently when a copy misses it. A missed table name in the claim
UPDATE matches zero rows, so `claimed !== 1` returns and the winner is
simply never paid, with nothing logged. A missed pair-limit count computes
one game's win-trading cap from the other game's history. None of it
surfaces as an error anyone can act on; it surfaces as "why am I not
getting points". persistAndAward is also exactly the code that gets patched
when a farm is found, so two copies guarantees a window where only one is
fixed, opening precisely when a new game invites new abuse.

So createDuelRuntime(spec) now holds the body and each game is a short
spec. Verified as a pure parameterization: stripping comments and
indentation, every differing executable line is a spec substitution plus
the four deliberate changes below. The full suite is unchanged at 322/323
(the one failure, monsterBattleEngine's attack-vs-attack KO, predates this
work and is untouched).

assertSpec() checks at boot the two things no type can: that the delegate
and the SQL table name describe the same table, and that the game is listed
in COMBAT_POOL_GAME_NAMES -- omitting it charges wins to both daily pools.

Four deliberate changes beyond the rename:

  * The per-pair daily award limit now spans every duel game rather than
    each game counting only its own table. The shared cap already bounds
    what one player earns; the pair limit exists to force a farmer to find
    distinct partners, and a per-game budget defeats that -- the same two
    accounts would trade their quota in RPS and then again in Pokemon.
  * Emits go through io.local. socket-server installs the Redis adapter, so
    a plain io.to() publishes each payload to Redis and reads it back only
    to discard it -- about 48 wasted round-trips per best-of-7 match. State
    here is single-process by design, so local is the correct path.
  * The intermission between rounds is per-result and shipped as an
    absolute nextRoundAt, like deadlineAt already is. A game that narrates
    a round needs longer to read than one that flashes a banner, and a
    client guessing at the duration drifts against the server's timer.
  * A self-match pays nothing. Unreachable today, but it is the only
    structural barrier if a bot seat ever reaches this path, and without it
    such a match falls through to same_device suppression.

The schema adds PokemonBattleMatch and PokemonBattleTrainer. The match
table carries a smaller index set than EdRpsMatch, which has three that
earn nothing: two partial indexes strictly supersede their plain
counterparts, and a bare endedAt index is never chosen because almost every
row satisfies the predicate. Each one dropped is a B-tree write saved on
the hot path, inside the award transaction under an advisory lock.
whoLeftUserId gains the index it lacks there -- it carries ON DELETE SET
NULL, so the dissolve worker would scan the whole table per deleted user.
Round history is stored positionally rather than as JSONB objects, which
repeat every key name in every element, and uploaded art lives in one JSON
map instead of eleven more columns on a table that already has 256.
